d/prune-unused-deps: include src/bootstrap
it's no longer part of the main workspace upstream, so we need to handle
both Cargo.lock files. to make the whole pruning similar (and since a
third Cargo.lock file for rust-analyzer will be added as well), switch
to using `cargo update --offline` for regenerating Cargo.lock, after
having confirmed that the resulting pruning is identical with both
invocations.

for this one upstream update, debian/prune-unused-deps needs to be run
outside of the git worktree instead of in a subdir, else cargo will
pickup the checked-in, unpatched 1.63 workspace Cargo.toml where
src/bootstrap was still a member, and fail because members of that
workspace are excluded via d/copyright, but the references are only gone
(and the Cargo.toml functional) *after* patching.

Signed-off-by: Fabian Grünbichler <debian@fabian.gruenbichler.email>